% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildUrl.R
\name{buildUrl}
\alias{buildUrl}
\title{Build url as per user input
Accepts user input for console name and timeframe}
\usage{
buildUrl()
}
\value{
url
}
\description{
Build url as per user input
Accepts user input for console name and timeframe
}
