% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnePlatformToRuleThemAll.R
\name{OnePlatformToRuleThemAll}
\alias{OnePlatformToRuleThemAll}
\title{Collects top games for each of the main consoles - ps4, xboxone, pc and switch
and returns games that are exclusive to each platform and summary stats for each
list of games by platform}
\usage{
OnePlatformToRuleThemAll()
}
\value{
list of dataframes - 
element 1 is a summary statistics dataframe, 
element 2 is a dataframe with ps4 exclusive games, 
element 3 is a dataframe with pc exclusive games, 
element 4 is a dataframe with switch exclusive games, 
element 5 is a dataframe with xboxone exclusive games
}
\description{
Collects top games for each of the main consoles - ps4, xboxone, pc and switch
and returns games that are exclusive to each platform and summary stats for each
list of games by platform
}
