% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readResponse.R
\name{readResponse}
\alias{readResponse}
\title{Writing function to obtain user input...}
\usage{
readResponse(p = "Enter Value:")
}
\arguments{
\item{p}{text prompt to guide user}
}
\value{
returns response entered by user
}
\description{
Writing function to obtain user input...
}
