% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{scraper}
\alias{scraper}
\title{Scrape webpage to return table with details on top games}
\usage{

  scraper(url = "http://www.metacritic.com/browse/games/score/metascore/90day/all/filtered?sort=desc")
}
\arguments{
\item{url}{metacritic top game list URL}
}
\value{
dataframe with top game data
}
\description{
Scrape webpage to return table with details on top games
}
