% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topgamesummary.R
\name{topgamesummary}
\alias{topgamesummary}
\title{Scrapes brief description of top 3 games for console specified by user from metacritic,
and formats collected data into dataframe}
\usage{

  topgamesummary(url = "http://www.metacritic.com/browse/games/score/metascore/all/ps4/filtered")
}
\arguments{
\item{url}{metacritic top games table url}
}
\value{
table with descriptions of top 3 games
}
\description{
Scrapes brief description of top 3 games for console specified by user from metacritic,
and formats collected data into dataframe
}
