% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualization}
\alias{visualization}
\title{Presents histograms and tables of data for the user's analysis}
\usage{
visualization(x = OnePlatformToRuleThemAll())
}
\arguments{
\item{x}{A list of five dataframes}
}
\value{
Prints histograms and tables of data
}
\description{
Presents histograms and tables of data for the user's analysis
}
